Moved thumbnail logic to a separate file
